chore: add comp for staff mode
